unittests, Models covered and OrderManagement almost coverd
